Extract error classes to dedicated errors module
