+++ src/fatsecret/__init__.py
@@ -1,6 +1,12 @@
-from .fatsecret import ApplicationError as ApplicationError
-from .fatsecret import AuthenticationError as AuthenticationError
-from .fatsecret import BaseFatsecretError as BaseFatsecretError
-from .fatsecret import Fatsecret as Fatsecret
-from .fatsecret import GeneralError as GeneralError
-from .fatsecret import ParameterError as ParameterError
+from .errors import (ApplicationError, AuthenticationError, BaseFatsecretError,
+                     GeneralError, ParameterError)
+from .fatsecret import Fatsecret
+
+__all__ = [
+    "Fatsecret",
+    "BaseFatsecretError",
+    "GeneralError",
+    "AuthenticationError",
+    "ParameterError",
+    "ApplicationError",
+]
